feat: github oAuth 로그인 구현 / 로그인 시에만 write 접근 기능 추가 / 본인 글만 삭제하는 기능 추가.

+++ src/pages/api/delete.ts
@@ -1,17 +1,25 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
 import { NextApiRequest, NextApiResponse } from "next";
+import { getServerSession } from "next-auth";
+import { authOptions } from "./auth/[...nextauth]";
 
 const postDeleteHandler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const { _id } = JSON.parse(req.body);
+  const session = await getServerSession(req, res, authOptions);
+
+  if (!session) return res.status(403).json("접근 권한이 없습니다.");
 
   if (req.method === "DELETE") {
-    const db = (await connectDB).db("forum");
-    const result = await db
-      .collection("post")
-      .deleteOne({ _id: new ObjectId(_id) });
+    const { _id, author } = JSON.parse(req.body);
 
-    return res.status(200).json("삭제됨");
+    if (session.user?.email === author) {
+      const db = (await connectDB).db("forum");
+      const result = await db
+        .collection("post")
+        .deleteOne({ _id: new ObjectId(_id) });
+
+      return res.status(200).json("삭제됨");
+    }
   }
 
   return res.status(500).json("I fucked up");
